Change backup path
Emprically, this seems like the correct path, because we see the error:

INFO    rsync: mkdir "/home/montagu/~/starport/orderly/data" failed: No such file or directory (2)

when running backup, which suggests that somewhere within bb8's code it
is appending together /home/<username> with the path provided;
presumably this changed - we think around the same time that the db
backup was changed to run from annex itself.

This change means that if we ever update the db backup on annex, it will
likely break, but it also seems clear that we should update this to
something we understand sooner rather than later.